Dialogo actualizado
Intro y final hechos, contexto lo dejo como estaba
