+++ index.html
@@ -42,7 +42,10 @@
                     <li> Para interactuar con un elemento pulsa la tecla <i>espacio</i>
                         <img src="./src/assets/Instructions/space.png" class="instr-pic">
                     </li>
-                    <li> Para empezar un minijuego interactúa con cualquiera de los <b>3</b> portales </li>
+                    <li> Para saltarte un dialogo /escena pulsa la tecla <i>escape</i>
+                        <img src="./src/assets/Instructions/space.png" class="instr-pic">
+                    </li>
+                    <li> Para empezar un minijuego interactúa con el centro de cualquiera de los <b>3</b> portales </li>
                 </ul>
             </div>
         </div>
@@ -55,7 +58,7 @@
                 <h3 class="modal-title">Descripción</h3>
                 <p> Las brujas de Zugarramurdi te han capturado para un sacrificio humano! :o </p>
                 <p> Si quieres escapar tendrás que completar una serie de arduas pruebas para conseguir la llave de la entrada, siempre y cuando no mueras en el intento >:) </p>
-                <p>Con este juego pretendemos hacer un homenaje al folklore vasco y a la leyenda de las brujas de Zugarramurdi.</p>
+                <p> Por desgracia para ti, las pruebas no están disponibles desde el inicio de la partida. Quien sabe, alomejor encuentras la forma de acceder a ellos explorando la cueva...</p>
             </div>
         </div> 
     </section>
